RL78 stuff

+++ platforms/rl78-gcc/stdint.h
@@ -38,6 +38,6 @@ typedef signed short int16_t;
 typedef signed long int32_t;
 typedef signed long long int64_t;
 
-typedef unsigned long uintptr_t;
+typedef unsigned short uintptr_t;
 
 #endif
